expose GetFutureEvents in sqlite backend
